requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-aiplatform==1.71.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
python-dotenv==1.0.0
python-json-logger==2.0.7

